// src/types/permissions.ts
export enum Permission {
  CREATE_USER = 'CREATE_USER',
  VIEW_USER = 'VIEW_USER',
  EDIT_USER = 'EDIT_USER',
  DELETE_USER = 'DELETE_USER',
  CREATE_DEPARTMENT = 'CREATE_DEPARTMENT',
  VIEW_DEPARTMENT = 'VIEW_DEPARTMENT',
  EDIT_DEPARTMENT = 'EDIT_DEPARTMENT',
  DELETE_DEPARTMENT = 'DELETE_DEPARTMENT',
  CREATE_POSITION = 'CREATE_POSITION',
  VIEW_POSITION = 'VIEW_POSITION',
  EDIT_POSITION = 'EDIT_POSITION',
  DELETE_POSITION = 'DELETE_POSITION',
  CREATE_SIGNATURE = 'CREATE_SIGNATURE',
  VIEW_SIGNATURE = 'VIEW_SIGNATURE',
  EDIT_SIGNATURE = 'EDIT_SIGNATURE',
  DELETE_SIGNATURE = 'DELETE_SIGNATURE',
  MANAGE_REQUEST_SIGN = 'MANAGE_REQUEST_SIGN',
  MANAGE_FILE_SIGN = 'MANAGE_FILE_SIGN',
  VIEW_ROLE = 'VIEW_ROLE',
  CREATE_ROLE = 'CREATE_ROLE',
  EDIT_ROLE = 'EDIT_ROLE',
  DELETE_ROLE = 'DELETE_ROLE',

  EDIT_WORKSPACE = 'EDIT_WORKSPACE',
  DELETE_WORKSPACE = 'DELETE_WORKSPACE',
  ADD_MEMBER = 'ADD_MEMBER',
  DELETE_MEMBER = 'DELETE_MEMBER',
  CREATE_STAGE = 'CREATE_STAGE',
  EDIT_STAGE = 'EDIT_STAGE',
  VIEW_LAYOUT_CHART = 'VIEW_LAYOUT_CHART',
  VIEW_LAYOUT_DOCUMENT = 'VIEW_LAYOUT_DOCUMENT',
  VIEW_LAYOUT_LIST = 'VIEW_LAYOUT_LIST',
  VIEW_LAYOUT_TIMELINE = 'VIEW_LAYOUT_TIMELINE',
  VIEW_ALL_TASK = 'VIEW_ALL_TASK',
  CREATE_TASK = 'CREATE_TASK',
  DELETE_TASK = 'DELETE_TASK',
  DELETE_STAGE = 'DELETE_STAGE',
}

export enum EWorkspacePermission {
  VIEW_ROLE = 'VIEW_ROLE',
  CREATE_ROLE = 'CREATE_ROLE',
  EDIT_ROLE = 'EDIT_ROLE',
  DELETE_ROLE = 'DELETE_ROLE',

  EDIT_WORKSPACE = 'EDIT_WORKSPACE',
  DELETE_WORKSPACE = 'DELETE_WORKSPACE',
  ADD_MEMBER = 'ADD_MEMBER',
  DELETE_MEMBER = 'DELETE_MEMBER',

  CREATE_STAGE = 'CREATE_STAGE',
  EDIT_STAGE = 'EDIT_STAGE',
  DELETE_STAGE = 'DELETE_STAGE',

  VIEW_LAYOUT_CHART = 'VIEW_LAYOUT_CHART',
  VIEW_LAYOUT_DOCUMENT = 'VIEW_LAYOUT_DOCUMENT',
  VIEW_LAYOUT_LIST = 'VIEW_LAYOUT_LIST',
  VIEW_LAYOUT_TIMELINE = 'VIEW_LAYOUT_TIMELINE',

  VIEW_ALL_TASK = 'VIEW_ALL_TASK',
  CREATE_TASK = 'CREATE_TASK',
  DELETE_TASK = 'DELETE_TASK',
}